all of my TODO implemented except in removeMeal and editMeal

+++ src/services/consumer/favoriteMeal/removeFavService.ts
@@ -3,6 +3,7 @@
  */
 require('dotenv').config();
 import { createClient } from "../../../config/db";
+import { checkFavMealExists, getUserFavMeal } from "../../../validations/consumer/favoriteMeal/removeFavValidation";
 
 /**
  * Class responsible for the service that serves to remove one of the meals from the favorites of the authenticated user
@@ -16,6 +17,18 @@ export class RemoveFavService {
      */
     async execute(uId: string, favMealId:string) {
 
+        const mealIdExists = await checkFavMealExists(favMealId)
+
+        if(!mealIdExists){
+            throw new Error('FavMeal does not exists')
+        }
+
+        const userMealId = await getUserFavMeal(favMealId);
+
+        if(userMealId != uId){
+            throw new Error('Favmeal does not belong to you')
+        }
+        
         const removeFavDBClient = createClient();
         const query = await removeFavDBClient.query(`DELETE FROM favoritemeals 
                                             WHERE uid = $1 AND favoritemealid = $2`, [uId, favMealId])
